refactor(web): extract the drag render layers into useDragLayers
The ghost (dragContentSvg), the frozen backdrop (dragStatic), the per-frame
preview geometry (dragPreview), the live edge and resize overlays
(liveEdges/liveNode), the gesture-frozen committed pair, and the keyed
patch container move from SpatialEditor.tsx into use-drag-layers.ts as one
derivation hook — verbatim logic and comments, closure reads turned into an
explicit inputs object. The hook owns no React state (one ref caches
carried edge sides per gesture), so the editor-state-surface ledger is
untouched.

purity-guard's file-seam property now scans both files, and its window is
the call's paren-matched argument span instead of 300 characters: the fixed
window reached into the memo dependency arrays where fileSeamOptions also
appears, so deleting the spread from the liveNode render left the old scan
green (measured). With the span window the same mutation fails the guard;
restore is green.

SpatialEditor.tsx: 3364 -> 3016 lines across this and the previous commit.
Verified: spatial-editor jsdom suite 45 files / 487 tests; browser suites
live-drag, live-drag-side-tracking, resize-live, drag-ghost-embed,
drag-content-preview, worker-scene-responsiveness, drag-start-
responsiveness, snap, multi-resize, object-first-connect, connect-preview
(42 tests) all pass unchanged.
