+++ apps/web/src/components/spatial-editor/SpatialEditor.tsx
@@ -59,24 +59,19 @@
  */
 
 import type {
-  EdgeSides,
   MeasureText,
   ResolvedReference,
   SpatialPresetKey,
-  TextMetrics,
 } from '@kamiazya/whiteboard-canvas-render'
 import {
   BODY_FONT_SIZE_PX,
   BODY_LINE_HEIGHT_PX,
   edgeLabelAnchor,
   flattenDrawnEdgePath,
-  layoutSpatialEdges,
-  renderSceneToSvg,
   SPATIAL_DARK_PALETTE,
   SPATIAL_LIGHT_PALETTE,
   SPATIAL_THEME_FONT_FAMILY,
   SPATIAL_THEME_GEOMETRY,
-  sceneBounds,
 } from '@kamiazya/whiteboard-canvas-render'
 import { createBrowserMeasureText } from '@kamiazya/whiteboard-canvas-viewer'
 import type { SpatialCanvas, SpatialNode } from '@kamiazya/whiteboard-model'
@@ -95,7 +90,6 @@ import { writeLastTool } from '@/lib/initial-tool'
 import type { ResolvedTheme } from '../../hooks/useThemeMode.js'
 import { parseClipboardText } from '../../lib/clipboard-fragment.js'
 import { hapticTick } from '../../lib/haptics.js'
-import { useKeyedSvg } from '../../lib/use-keyed-svg.js'
 import type { BoxMove } from './align.js'
 import {
   CanvasContextMenu,
@@ -109,7 +103,7 @@ import { applyCommand } from './commands.js'
 import { CREATION_LABELS } from './creation-labels.js'
 import { DocumentPickerDialog, type FileRefOption } from './DocumentPickerDialog.js'
 import { DragPreviewLayer } from './DragPreviewLayer.js'
-import { computeDragPreview, isInFlightGesture } from './drag-preview.js'
+import { isInFlightGesture } from './drag-preview.js'
 import { createEditorAppearance, editorTextFill } from './editor-appearance.js'
 import { FacetFormPanel } from './facet-widgets/FacetFormPanel.js'
 import { isFollowableUrl } from './followable-url.js'
@@ -126,14 +120,7 @@ import {
 } from './geometry.js'
 import { snapGesturePoint } from './gesture-snap.js'
 import { describeTarget, gestureTrace } from './gesture-trace.js'
-import {
-  type CarriedSideCache,
-  canReuseCarriedSides,
-  carriedByGesture,
-  carriedSideCacheKey,
-  frozenSidesOf,
-  liveNodesFor,
-} from './gesture-view.js'
+import { carriedByGesture } from './gesture-view.js'
 import type { GestureState } from './gestures.js'
 import { createIdleState, NEW_NODE_HEIGHT, NEW_NODE_WIDTH, reduceGesture } from './gestures.js'
 import { LinkEmbedLayer } from './LinkEmbedLayer.js'
@@ -152,7 +139,7 @@ import {
 } from './navigation.js'
 import { PendingCutChip } from './PendingCutChip.js'
 import { SelectionOverlay } from './SelectionOverlay.js'
-import { renderCanvasToSvg, requiredTextNodeHeight } from './scene-render.js'
+import { requiredTextNodeHeight } from './scene-render.js'
 import { renderedCanvasKeyed } from './scene-render-core.js'
 import {
   EMPTY_SELECTION,
@@ -169,9 +156,9 @@ import {
   ToolPalette,
 } from './ToolPalette.js'
 import { useClipboardActions } from './use-clipboard-actions.js'
+import { useDragLayers } from './use-drag-layers.js'
 import { MINIMAP_MIN_ROOT_WIDTH_PX, useEditorMeasurements } from './use-editor-measurements.js'
 import { EXPAND_MIN_H, EXPAND_MIN_W, useFileSeamScene } from './use-file-seam-scene.js'
-import { useGestureCaptured } from './use-gesture-captured.js'
 import { useNativeCanvasListeners } from './use-native-canvas-listeners.js'
 import { useNodeCreation } from './use-node-creation.js'
 import { useWorkerScene } from './use-worker-scene.js'
@@ -820,137 +807,28 @@ export const SpatialEditor = forwardRef<SpatialEditorHandle, SpatialEditorProps>
       // isLocked closes over lockedNodeIds/lockEnabled, both listed here.
     }, [lockEnabled, lockedNodeIds, selectedId, extraIds, gestureState])
 
-    /**
-     * The dragged node's own content, rendered ONCE per drag (the reducer's
-     * pointermove passthrough returns the same state reference, so this memo
-     * holds for the whole gesture; a single-node render costs ~0.4ms).
-     * Per-frame motion is then a pure CSS transform in DragPreviewLayer —
-     * the full-canvas render stays untouched during the drag.
-     */
-    const dragContentSvg = useMemo(() => {
-      if (gestureState.kind !== 'moving') return undefined
-      const carried = carriedByGesture(canvas, gestureState, extraIds, isLocked)
-      const nodes = canvas.nodes.filter((n) => carried.has(n.id))
-      if (nodes.length === 0) return undefined
-      // Same embed options as the committed scene: a ghost that drops an
-      // expanded miniature back to a bare card mid-drag reads as data loss.
-      const rendered = renderCanvasToSvg(
-        { nodes, edges: [] },
-        {
-          measure: resolvedMeasure,
-          theme,
-          ...fileSeamOptions,
-        },
-      )
-      return {
-        svg: rendered.svg,
-        originX: gestureState.startX - rendered.bounds.x,
-        originY: gestureState.startY - rendered.bounds.y,
-      }
-      // isLocked closes over lockedNodeIds/lockEnabled, both listed.
-    }, [
-      gestureState,
-      canvas,
-      extraIds,
-      lockEnabled,
-      lockedNodeIds,
-      resolvedMeasure,
-      theme,
-      fileSeamOptions,
-    ])
-
-    /**
-     * The scene WITHOUT everything the drag layers draw live: carried
-     * nodes travel as the ghost, and EVERY edge re-routes per frame in
-     * the live-edges layer — a bystander edge is excluded too, because
-     * the moving node entering or leaving its path changes its route and
-     * its line jumps, and a frozen copy would disagree with the drop
-     * result. Rendered ONCE per drag (gestureState is reference-stable
-     * across pointermoves), so per-frame cost stays with the small layers.
-     * The returned `measure` memoizes per drag: edge labels measure on the
-     * first live frame and every later frame re-places the cached metrics,
-     * keeping pointermoves free of text measurement.
-     * ponytail: the backdrop render here is ~21ms at 45 nodes (the anchor
-     * pass, formerly ~7x that, now arrives with the committed scene); if
-     * start jank reappears on much larger documents, the next rung is
-     * reusing the committed scene graph for the backdrop instead of
-     * re-rendering — drop the carried node runs, truncate at the first
-     * edge, re-render the remainder (composeNode is per-node pure, so the
-     * prefix equivalence holds while the backdrop stays edge-free).
-     */
-    // The committed layout, FROZEN at gesture start. The worker's next
-    // reply may land mid-gesture, and BOTH halves matter: the anchors are
-    // the points bystander edges are pinned to, and the scene is what
-    // decides which edges are pin-eligible at all (frozenSidesOf) — a
-    // swapped scene silently un-pins every bystander even with the anchors
-    // held, which is the same re-fraction by another door.
-    const committedPair = useMemo(() => ({ scene, anchors }), [scene, anchors])
-    const gestureCommitted = useGestureCaptured(
-      gestureState.kind === 'moving' ||
-        gestureState.kind === 'resizing' ||
-        gestureState.kind === 'connecting',
-      committedPair,
-    )
-    // Last optimized sides for the gesture's carried edges (see liveEdges).
-    const carriedSideCacheRef = useRef<CarriedSideCache | null>(null)
-    useEffect(() => {
-      if (gestureState.kind !== 'moving' && gestureState.kind !== 'resizing') {
-        carriedSideCacheRef.current = null
-      }
-    }, [gestureState.kind])
-
-    const dragStatic = useMemo(() => {
-      if (gestureState.kind !== 'moving' && gestureState.kind !== 'resizing') return undefined
-      const carried = carriedByGesture(canvas, gestureState, extraIds, isLocked)
-      const base: SpatialCanvas = {
-        ...canvas,
-        nodes: canvas.nodes.filter((n) => !carried.has(n.id)),
-        edges: [],
-      }
-      const rendered = renderCanvasToSvg(base, {
-        measure: resolvedMeasure,
+    // The drag/resize/connect render layers (ghost, backdrop, live edges,
+    // live resize, preview geometry, and the committed surface's mount-once
+    // patch container) — one derivation hook, no state of its own. See
+    // use-drag-layers.ts for the per-gesture vs per-frame split.
+    const { dragContentSvg, dragStatic, dragPreview, liveEdges, liveNode, canvasContentRef } =
+      useDragLayers({
+        gestureState,
+        canvas,
+        extraIds,
+        isLocked,
+        lockEnabled,
+        lockedNodeIds,
+        resolvedMeasure,
         theme,
-        ...fileSeamOptions,
+        fileSeamOptions,
+        scene,
+        anchors,
+        keyed,
+        boxes,
+        selectableBoxes,
+        livePoint,
       })
-      const metricsCache = new Map<string, TextMetrics>()
-      const measure: MeasureText = (text, font) => {
-        const key = `${font.family}|${font.weight}|${font.style}|${font.sizePx}\u0000${text}`
-        const hit = metricsCache.get(key)
-        if (hit !== undefined) return hit
-        const metrics = resolvedMeasure(text, font)
-        metricsCache.set(key, metrics)
-        return metrics
-      }
-      // The committed anchor state: liveEdges pins bystander edges to these
-      // exact points so a carried edge joining their (node, side) group
-      // cannot re-fraction them mid-drag. Taken from the committed scene's
-      // OWN layout rather than re-run here — the anchor pass is the most
-      // expensive step of a layout (measured: ~7x the backdrop render), and
-      // these are also the anchors the pixels on screen were routed with,
-      // which a fresh pass over a newer canvas is not.
-      return {
-        carried,
-        keyed: renderedCanvasKeyed(rendered),
-        bounds: rendered.bounds,
-        measure,
-        committedAnchors: gestureCommitted.anchors,
-      }
-      // isLocked closes over lockedNodeIds/lockEnabled, both listed.
-    }, [
-      gestureState,
-      canvas,
-      gestureCommitted,
-      extraIds,
-      lockEnabled,
-      lockedNodeIds,
-      resolvedMeasure,
-      theme,
-      fileSeamOptions,
-    ])
-    // The committed surface's mount-once container (see the JSX below):
-    // during a gesture it patches to the drag backdrop, on drop back to the
-    // committed render — both through the same keyed reconciliation.
-    const canvasContentRef = useKeyedSvg(dragStatic?.keyed ?? keyed)
 
     useImperativeHandle(
       forwardedRef,
@@ -998,160 +876,6 @@ export const SpatialEditor = forwardRef<SpatialEditorHandle, SpatialEditorProps>
       [selectionMembers],
     )
     const isMultiSelection = selectionMembers.length > 1
-
-    /**
-     * The in-flight preview geometry, derived per frame from the gesture's own
-     * start snapshot plus the live pointer. Move/resize never read `canvas`
-     * (see drag-preview.ts for why, and for the single-source
-     * `resizeBoxByDelta` guarantee it documents); the connecting branch does —
-     * it routes the prospective edge through the committed producer, a few
-     * routeEdge calls per frame.
-     */
-    const dragPreview = useMemo(() => {
-      // Existing edges keep their committed sides while a connect gesture
-      // is in flight — same freeze the live drag overlay applies, so the
-      // canvas around the pointer stays still and pointer frames skip the
-      // crossing-optimization loop. The prospective edge itself derives
-      // fresh each frame.
-      const frozenEdgeSides = frozenSidesOf(gestureCommitted.scene)
-      return computeDragPreview(gestureState, boxes, livePoint, {
-        canvas,
-        selectableBoxes,
-        frozenEdgeSides,
-      })
-    }, [gestureState, livePoint, boxes, canvas, selectableBoxes, gestureCommitted])
-
-    /**
-     * EVERY edge, re-composed against the ghost's snapped live position and
-     * rendered as an overlay — the per-frame half of live drag rendering.
-     * Goes through canvas-render's `layoutSpatialEdges`, the same producer
-     * the committed render uses, so routing detours around the moving node,
-     * line jumps, and label placement all match the drop result exactly
-     * (one producer per geometry). Per pointermove this is edge routing
-     * plus a small serialization; text measurement is absorbed by
-     * `dragStatic.measure`'s per-drag cache.
-     */
-    const liveEdges = useMemo(() => {
-      if (
-        (gestureState.kind !== 'moving' && gestureState.kind !== 'resizing') ||
-        dragPreview === undefined ||
-        dragPreview.kind !== 'box' ||
-        dragStatic === undefined ||
-        canvas.edges.length === 0
-      ) {
-        return undefined
-      }
-      const liveNodes = [...liveNodesFor(canvas, gestureState, dragPreview.box, dragStatic.carried)]
-      // BYSTANDER sides stay frozen at their committed choices for the whole
-      // gesture: re-optimizing them per frame would let unrelated routes
-      // flip sides mid-drag. Edges attached to a CARRIED node re-optimize
-      // through the same side optimizer the committed render uses (so the
-      // drop cannot re-side an edge the preview never showed that way) —
-      // but only once per CARRIED_RESIDE_STEP_PX of travel: the optimizer's
-      // trial loop costs ~8-14ms and a side decision rarely changes within
-      // a few pixels, so in-between frames reuse the cached sides as a full
-      // override map, which skips the optimizer entirely.
-      const carried = dragStatic.carried
-      const carriedEdgeIds = new Set(
-        canvas.edges
-          .filter((edge) => carried.has(edge.fromNode) || carried.has(edge.toNode))
-          .map((edge) => edge.id),
-      )
-      const frozenSides = new Map(
-        [...frozenSidesOf(gestureCommitted.scene)]
-          .filter(([id]) => !carriedEdgeIds.has(id))
-          .map(([id, pair]) => {
-            const pin = dragStatic.committedAnchors.get(id)
-            return [
-              id,
-              {
-                ...pair,
-                from: pin?.from,
-                fromLaneDepth: pin?.fromLaneDepth,
-                to: pin?.to,
-                toLaneDepth: pin?.toLaneDepth,
-              },
-            ] as const
-          }),
-      )
-      const cacheKey = carriedSideCacheKey(carriedEdgeIds)
-      const cache = carriedSideCacheRef.current
-      const reuse = canReuseCarriedSides(cache, cacheKey, dragPreview.box.x, dragPreview.box.y)
-      const overrides =
-        reuse && cache !== null ? new Map([...frozenSides, ...cache.sides]) : frozenSides
-      const nodes = layoutSpatialEdges(
-        { ...canvas, nodes: liveNodes },
-        {
-          measure: dragStatic.measure,
-          appearance: createEditorAppearance(theme),
-          edgeSideOverrides: overrides,
-        },
-      )
-      if (!reuse) {
-        const sides = new Map<string, EdgeSides>()
-        for (const node of nodes) {
-          if (node.kind === 'edge' && carriedEdgeIds.has(node.id)) {
-            sides.set(node.id, { fromSide: node.fromSide, toSide: node.toSide })
-          }
-        }
-        carriedSideCacheRef.current = {
-          key: cacheKey,
-          anchorX: dragPreview.box.x,
-          anchorY: dragPreview.box.y,
-          sides,
-        }
-      }
-      const liveBounds = sceneBounds({ nodes })
-      return {
-        svg: renderSceneToSvg(
-          { nodes },
-          { width: liveBounds.w, height: liveBounds.h, viewBox: liveBounds },
-        ),
-        bounds: liveBounds,
-      }
-    }, [gestureState, dragPreview, dragStatic, canvas, theme, gestureCommitted])
-
-    /**
-     * The resized node's own content, re-rendered at its PREVIEW size each
-     * frame — a resize changes geometry, so the move ghost's render-once-
-     * transform-per-frame trick cannot apply. Affordable because it is one
-     * node (~0.4ms) and `dragStatic.measure` memoizes text metrics for the
-     * gesture: the first frame warms the cache and later frames re-wrap
-     * with zero new measure calls.
-     *
-     * File-node LOD (card vs inline embed) deliberately stays at its
-     * COMMITTED decision for the whole gesture — the same freeze-then-
-     * settle rule edge sides follow: a mid-gesture card/embed swap is
-     * exactly the kind of flicker the freeze exists to prevent, and the
-     * expansion hysteresis is stateful over the committed canvas. The
-     * crossing of a size threshold takes effect on release.
-     */
-    const liveNode = useMemo(() => {
-      if (
-        gestureState.kind !== 'resizing' ||
-        dragPreview === undefined ||
-        dragPreview.kind !== 'box' ||
-        dragStatic === undefined
-      ) {
-        return undefined
-      }
-      const resized = liveNodesFor(
-        canvas,
-        gestureState,
-        dragPreview.box,
-        new Set([gestureState.nodeId]),
-      ).find((n) => n.id === gestureState.nodeId)
-      if (resized === undefined) return undefined
-      const rendered = renderCanvasToSvg(
-        { nodes: [resized], edges: [] },
-        {
-          measure: dragStatic.measure,
-          theme,
-          ...fileSeamOptions,
-        },
-      )
-      return { svg: rendered.svg, bounds: rendered.bounds }
-    }, [gestureState, dragPreview, dragStatic, canvas, theme, fileSeamOptions])
 
     /**
      * How far a snap guide extends, in canvas space: across all content plus
